+++ src/item.cpp
@@ -26,7 +26,7 @@ Block::Block(Game* game, BufferedBlock* d) : Item(game), data(d) {
 
     if (data->sprites != 0) {
         H2DE_Timeline* t = H2DE_CreateTimeline(engine, 500, LINEAR, [this](float blend) {
-            this->currentSprite = std::min((int)floor(blend * this->data->sprites), this->data->sprites - 1) + 1;
+            this->currentSprite = std::min((int)floor(lerp(0, this->data->sprites, blend)), this->data->sprites - 1) + 1;
         }, NULL, -1);
 
         H2DE_AddTimelineToManager(tm, t);
@@ -42,48 +42,56 @@ Block::~Block() {
 // UPDATE
 void Block::update() {
     static H2DE_Engine* engine = game->getEngine();
+    static GameData* gameData = game->getData();
     Level* level = game->getLevel();
     float camPosX = game->getCamera()->getPos().x;
 
-    float maxCamX = camPosX + BLOCKS_ON_WIDTH + 0.3f;
+    static float camWidth = gameData->sizes->camera.w;
+    static float camXremain = gameData->sizes->cameraXremain;
+    static float blockEffectPadding = gameData->sizes->blockEffectPadding;
+    static float moveOffset = gameData->offsets->blockEffectMove;
+
+    float maxCamX = camPosX + camWidth;
     bool onScreen = (maxCamX > data->pos.x);
 
     BlockEffect levelBlockEffect = level->getBlockEffect();
 
     if (onScreen) {
-        bool onMaxXedge = (data->pos.x >= maxCamX - 2.5f);
-        bool onMinXedge = (data->pos.x - 2.5f <= camPosX);
+        bool onMaxXedge = (data->pos.x >= maxCamX - blockEffectPadding);
+        bool onMinXedge = (data->pos.x - blockEffectPadding <= camPosX);
 
-        float blend, min = 0.0f, max = 0.0f;
+        float blend, minOffset = 0.0f, maxOffset = 0.0f, minSize = 0.0f, maxSize = 0.0f;
         if (onMaxXedge) {
             if (!effect.has_value()) effect = levelBlockEffect;
 
-            float dist = maxCamX - data->pos.x - 0.5f;
-            blend = std::clamp(dist / 2.0f, 0.0f, 1.0f);
-            min = 2.0f;
-            max = 0.0f;
-            textureOpacity = lerp(0, 255, blend);
+            float dist = maxCamX - data->pos.x - camXremain;
+            blend = std::clamp(dist / (blockEffectPadding - 0.5f), 0.0f, 1.0f);
+            textureOpacity = lerp(SDL_MIN_UINT8, SDL_MAX_UINT8, blend);
+            minOffset = moveOffset;
+            maxOffset = 0.0f;
+            minSize = 0.0f;
+            maxSize = 1.0f;
 
         } else if (onMinXedge) {
             if (!effect.has_value()) effect = levelBlockEffect;
 
-            float dist = camPosX + 1.5f - data->pos.x;
-            blend = std::clamp(dist / 2.0f, 0.0f, 1.0f);
-            textureOpacity = lerp(255, 0, blend);
-            min = 0.0f;
-            max = 2.0f;
+            float dist = camPosX + blockEffectPadding - 1.0f - data->pos.x;
+            blend = std::clamp(dist / (blockEffectPadding - 0.5f), 0.0f, 1.0f);
+            textureOpacity = lerp(SDL_MAX_UINT8, SDL_MIN_UINT8, blend);
+            minOffset = 0.0f;
+            maxOffset = moveOffset;
+            minSize = 1.0f;
+            maxSize = 0.0f;
 
-        } else blend = 1.0f;
+        } else effect = std::nullopt;
         
         if (effect.has_value()) switch (effect.value()) {
-             case FROM_TOP: textureOffset = { 0.0f, lerp(min, max, blend) }; break;
-             case FROM_BOTTOM: textureOffset = { 0.0f, lerp(-min, -max, blend) }; break;
-             case FROM_LEFT: textureOffset = { lerp(-min, -max, blend), 0.0f }; break;
-             case FROM_RIGHT: textureOffset = { lerp(min, max, blend), 0.0f }; break;
-             case SCALE: textureScale = 0.0f; break;
+             case FROM_TOP: textureOffset = { 0.0f, lerp(minOffset, maxOffset, blend) }; break;
+             case FROM_BOTTOM: textureOffset = { 0.0f, lerp(-minOffset, -maxOffset, blend) }; break;
+             case FROM_LEFT: textureOffset = { lerp(-minOffset, -maxOffset, blend), 0.0f }; break;
+             case FROM_RIGHT: textureOffset = { lerp(minOffset, maxOffset, blend), 0.0f }; break;
+             case SCALE: textureScale = lerp(minSize, maxSize, blend); break;
         }
-
-        if (!onMaxXedge && !onMinXedge) effect = std::nullopt;
     } else textureOpacity = 0;
 }
 
@@ -92,40 +100,63 @@ void Block::render() {
     LevelPos camPos = game->getCamera()->getPos();
     GameState gameState = game->getState();
 
-    if (!pickedUp) if (data->pos.x + data->textureSize.w >= camPos.x && data->pos.x - data->textureSize.w <= camPos.x + BLOCKS_ON_WIDTH + 1) {
+    if (!pickedUp) if (data->pos.x + data->textureSize.w >= camPos.x && data->pos.x - data->textureSize.w <= camPos.x + BLOCKS_ON_WIDTH + 1) { // refaire
         if (gameState.main != LEVEL_PAUSE) H2DE_TickTimelineManager(tm);
         renderTexture();
+        if (data->glowID != 0) renderGlow();
         if (game->getMegahack()->getHack("show-hitboxes")->active) renderHitbox();
     }
 }
 
 void Block::renderTexture() {
     static H2DE_Engine* engine = game->getEngine();
-    static GameData* gameData = game->getData();
     static Calculator* calculator = game->getCalculator();
     Level* level = game->getLevel();
-    LevelPos playerPos = level->getPlayer()->getPos();
     LevelPos camPos = game->getCamera()->getPos();
 
-    bool textureOnCamX = (data->pos.x + data->textureOffset.x + data->textureSize.w > camPos.x) && (data->pos.x + data->textureOffset.x < camPos.x + BLOCKS_ON_WIDTH + 1);
-    if (textureOnCamX) {
-        std::string tex = data->texture.substr(0, data->texture.size() - 4).append(((data->sprites != 0) ? std::string("-").append(std::to_string(currentSprite)) : "")).append(".png");
-        H2DE_Size absTexOri = calculator->convertToPx(LevelSize{ data->rotationOrigin.x, data->rotationOrigin.y });
-        H2DE_RGB rgb = static_cast<H2DE_RGB>(level->getData()->colors[data->colorID]);
-        rgb.a = textureOpacity;
+    std::string tex = data->texture.substr(0, data->texture.size() - 4).append(((data->sprites != 0) ? std::string("-").append(std::to_string(currentSprite)) : "")).append(".png");
+    H2DE_Size absTexOrigin = calculator->convertToPx(LevelSize{ data->texOrigin.x, data->texOrigin.y });
+    H2DE_RGB rgb = static_cast<H2DE_RGB>(level->getData()->colors[data->colorID]);
+    rgb.a = textureOpacity;
 
-        H2DE_GraphicObject* texture = H2DE_CreateGraphicObject();
-        texture->type = IMAGE;
-        texture->texture = tex;
-        texture->pos = calculator->convertToPx({ data->pos.x + data->textureOffset.x + textureOffset.x, data->pos.y + data->textureOffset.y + textureOffset.y }, data->textureSize, false, false);
-        texture->size = calculator->convertToPx(data->textureSize);
-        texture->rotation = data->rotation;
-        texture->rotationOrigin = { absTexOri.w, absTexOri.h };
-        texture->flip = data->flip;
-        texture->rgb = rgb;
-        texture->index = data->zIndex->getIndex();
-        H2DE_AddGraphicObject(engine, texture);
-    }
+    H2DE_GraphicObject* texture = H2DE_CreateGraphicObject();
+    texture->type = IMAGE;
+    texture->texture = tex;
+    texture->pos = calculator->convertToPx({ data->pos.x + data->textureOffset.x + textureOffset.x, data->pos.y + data->textureOffset.y + textureOffset.y }, data->textureSize, false, false);
+    texture->size = calculator->convertToPx(data->textureSize);
+    texture->rotation = data->rotation;
+    texture->rotationOrigin = { absTexOrigin.w, absTexOrigin.h };
+    texture->scale = { textureScale, textureScale };
+    texture->scaleOrigin = { absTexOrigin.w, absTexOrigin.h };
+    texture->flip = data->flip;
+    texture->rgb = rgb;
+    texture->index = data->zIndex->getIndex();
+    H2DE_AddGraphicObject(engine, texture);
+}
+
+void Block::renderGlow() {
+    static H2DE_Engine* engine = game->getEngine();
+    static GameData* gameData = game->getData();
+    static Calculator* calculator = game->getCalculator();
+
+    static float glowOpacity = gameData->colors->glowOpacity;
+    H2DE_Size absGlowOrigin = calculator->convertToPx(LevelSize{ data->glowOrigin.x, data->glowOrigin.y });
+    H2DE_RGB rgb = { 255, 255, 255, 255 }; // replace => pcol 1 or 2 (idk)
+    rgb.a = textureOpacity * glowOpacity;
+
+    H2DE_GraphicObject* glow = H2DE_CreateGraphicObject();
+    glow->type = IMAGE;
+    glow->texture = "glow-" + std::to_string(data->glowID) + ".png";
+    glow->pos = calculator->convertToPx({ data->pos.x + textureOffset.x + data->glowOffset.x, data->pos.y + textureOffset.y + data->glowOffset.y }, data->glowSize, false, false);
+    glow->size = calculator->convertToPx(data->glowSize);
+    glow->rotation = data->rotation;
+    glow->rotationOrigin = { absGlowOrigin.w, absGlowOrigin.h };
+    glow->scale = { textureScale, textureScale };
+    glow->scaleOrigin = { absGlowOrigin.w, absGlowOrigin.h };
+    glow->flip = data->flip;
+    glow->rgb = rgb;
+    glow->index = data->zIndex->getIndex() - 1;
+    H2DE_AddGraphicObject(engine, glow);
 }
 
 void Block::renderHitbox() {
@@ -134,8 +165,7 @@ void Block::renderHitbox() {
     static Calculator* calculator = game->getCalculator();
     LevelPos camPos = game->getCamera()->getPos();
 
-    bool hitboxOnCamX = (data->pos.x + data->hitboxOffset.x + data->hitboxSize.w > camPos.x) && (data->pos.x + data->hitboxOffset.x < camPos.x + BLOCKS_ON_WIDTH + 1);
-    if (hitboxOnCamX && data->type != DECORATION) {
+    if (data->type != DECORATION) {
         H2DE_Size absHitboxSize = calculator->convertToPx(data->hitboxSize);
         H2DE_Pos absHitboxOffset = calculator->convertToPx(data->hitboxOffset);
         H2DE_GraphicObject* hitbox = H2DE_CreateGraphicObject();
@@ -158,7 +188,7 @@ void Block::reset() {
     used = false;
     pickedUp = false;
 
-    textureOpacity = 255;
+    textureOpacity = SDL_MAX_UINT8;
     textureOffset = { 0.0f, 0.0f };
     textureScale = 1;
     effect = std::nullopt;
@@ -270,10 +300,10 @@ void Trigger::trigger() {
         H2DE_Timeline* t = H2DE_CreateTimeline(engine, data->ms, LINEAR, [this, level, sCol](float blend) {
             Color eCol = data->color;
             Color bCol = {
-                static_cast<Uint8>(sCol.r + blend * (eCol.r - sCol.r)),
-                static_cast<Uint8>(sCol.g + blend * (eCol.g - sCol.g)),
-                static_cast<Uint8>(sCol.b + blend * (eCol.b - sCol.b)),
-                static_cast<Uint8>(sCol.a + blend * (eCol.a - sCol.a)),
+                static_cast<Uint8>(lerp(sCol.r, eCol.r, blend)),
+                static_cast<Uint8>(lerp(sCol.g, eCol.g, blend)),
+                static_cast<Uint8>(lerp(sCol.b, eCol.b, blend)),
+                static_cast<Uint8>(lerp(sCol.a, eCol.a, blend)),
             };
 
             switch (data->type) {